requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
frozenlist==1.7.0
google-auth==2.40.3
h11==0.16.0
httptools==0.6.4
idna==3.10
multidict==6.6.3
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
